dev_requirements.txt
djangorestframework-stubs
pylint
pylint-django
pytest
pytest-cov
pytest-django
pytest-xdist
flake8